add TCP and CNAME checks to okdns
adds a check for querying the server over TLS as well as UDP, and a
check to make sure that the domain includes no CNAME records.

swaps the IsFailed func for a Success func that includes checking the
number of errors for each nameserver. text output includes errors
encountered when running queries.